Fix PWM calculations
Changed algorithm we're using to calculate the PWM duty cycle to reset the timer after pushing into each ring buffer to make it more robust. Modify scaling algorithm.

+++ drivers/current_sensor.h
@@ -6,4 +6,4 @@
 void current_sensor_init(const struct PWMConfig *pwm);
 
 // Returns the amperage from the current sensor
-int16_t current_sensor_read(void);
+int32_t current_sensor_read(void);
